requirement.txt
Django==4.1
Pillow==9.2.0
django-tinymce==3.4.0
beautifulsoup4==4.11.1
